document.rs and normal.rs test clean up
